+++ src/main/kotlin/io/github/ackeecz/danger/testing/junit/JUnitResultFileFinder.kt
@@ -1,13 +1,11 @@
 package io.github.ackeecz.danger.testing.junit
 
 import io.github.ackeecz.danger.testing.FileFinder
-import io.github.ackeecz.danger.testing.NoFilesFoundException
+import systems.danger.kotlin.sdk.DangerContext
 import java.io.File
 import java.nio.file.Path
 
-internal object JUnitResultFileFinder {
-
-    private const val RESULT_FILE_EXTENSION = "xml"
+internal class JUnitResultFileFinder(private val context: DangerContext) {
 
     fun findFiles(
         rootDirectoryPath: Path,
@@ -19,8 +17,15 @@ internal object JUnitResultFileFinder {
             reportFilesFolderPath = config.testResultsFolderPath,
             filePrefix = config.resultFileNamePrefix,
             fileExtension = RESULT_FILE_EXTENSION,
-        ).ifEmpty {
-            throw NoFilesFoundException("No test results found. Check your configuration.")
+        ).also {
+            if (it.isEmpty()) {
+                context.warn("No JUnit test results found")
+            }
         }
+    }
+
+    companion object {
+
+        private const val RESULT_FILE_EXTENSION = "xml"
     }
 }